peige/requirements.txt
Django>=2.1,<2.2
djangorestframework==3.9.1
gunicorn==19.7.1